Add BridgePanel component, moved logic from index to BridgePanel.
